phoenix_thighs/requirements.txt
pyautogui
opencv-python
numpy
pydirectinput
tf-centernet
pyserial